--- cmd/counter/main.go
@@ -1,32 +1,0 @@
-package main
-
-import (
-	"os"
-
-	"github.com/spf13/cobra"
-
-	"github.com/tendermint/tmlibs/cli"
-
-	"github.com/tendermint/basecoin/cmd/basecoin/commands"
-	"github.com/tendermint/basecoin/docs/guide/counter/plugins/counter"
-)
-
-func main() {
-	var RootCmd = &cobra.Command{
-		Use:   "counter",
-		Short: "demo plugin for basecoin",
-	}
-
-	// TODO: register the counter here
-	commands.Handler = counter.NewHandler("mycoin")
-
-	RootCmd.AddCommand(
-		commands.InitCmd,
-		commands.StartCmd,
-		commands.UnsafeResetAllCmd,
-		commands.VersionCmd,
-	)
-
-	cmd := cli.PrepareMainCmd(RootCmd, "CT", os.ExpandEnv("$HOME/.counter"))
-	cmd.Execute()
-}
