+++ cmd/gaia/main.go
@@ -1,0 +1,61 @@
+package main
+
+import (
+	"os"
+
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+	"github.com/tendermint/tmlibs/cli"
+	"github.com/tendermint/tmlibs/log"
+
+	"github.com/cosmos/gaia/modules/stake"
+
+	"github.com/tendermint/basecoin/cmd/basecoin/commands"
+)
+
+//nolint
+const (
+	defaultLogLevel = "error"
+	FlagLogLevel    = "log_level"
+)
+
+var (
+	logger = log.NewTMLogger(log.NewSyncWriter(os.Stdout)).With("module", "main")
+)
+
+// RootCmd - main node command
+var RootCmd = &cobra.Command{
+	Use:   "gaia",
+	Short: "The Cosmos Network delegation-game blockchain",
+	PersistentPreRunE: func(cmd *cobra.Command, args []string) (err error) {
+		level := viper.GetString(FlagLogLevel)
+		logger, err = tmflags.ParseLogLevel(level, logger, defaultLogLevel)
+		if err != nil {
+			return err
+		}
+		if viper.GetBool(cli.TraceFlag) {
+			logger = log.NewTracingLogger(logger)
+		}
+		return nil
+	},
+}
+
+func init() {
+	RootCmd.PersistentFlags().String(FlagLogLevel, defaultLogLevel, "Log level")
+}
+
+func main() {
+	// require all fees in mycoin - change this in your app!
+	commands.Handler = stake.NewHandler("mycoin")
+
+	RootCmd.AddCommand(
+		commands.InitCmd,
+		commands.StartCmd,
+		//commands.RelayCmd,
+		commands.UnsafeResetAllCmd,
+		commands.VersionCmd,
+	)
+
+	cmd := cli.PrepareMainCmd(RootCmd, "GA", os.ExpandEnv("$HOME/.cosmos-gaia"))
+	cmd.Execute()
+}
