+++ cmd/gaiacli/commands/query.go
