+++ cmd/gaiacli/main.go
@@ -21,20 +21,14 @@ import (
 	noncecmd "github.com/tendermint/basecoin/modules/nonce/commands"
 )
 
-// BaseCli represents the base command when called without any subcommands
-var BaseCli = &cobra.Command{
-	Use:   "countercli",
-	Short: "Light client for tendermint",
-	Long: `Basecli is an version of tmcli including custom logic to
-present a nice (not raw hex) interface to the basecoin blockchain structure.
-
-This is a useful tool, but also serves to demonstrate how one can configure
-tmcli to work for any custom abci app.
-`,
+// GaiaCli represents the base command when called without any subcommands
+var GaiaCli = &cobra.Command{
+	Use:   "gaiacli",
+	Short: "Client for Cosmos-Gaia blockchain",
 }
 
 func main() {
-	commands.AddBasicFlags(BaseCli)
+	commands.AddBasicFlags(GaiaCli)
 
 	// Prepare queries
 	proofs.RootCmd.AddCommand(
@@ -68,7 +62,7 @@ func main() {
 	)
 
 	// Set up the various commands to use
-	BaseCli.AddCommand(
+	GaiaCli.AddCommand(
 		commands.InitCmd,
 		commands.ResetCmd,
 		keycmd.RootCmd,
@@ -78,6 +72,6 @@ func main() {
 		proxy.RootCmd,
 	)
 
-	cmd := cli.PrepareMainCmd(BaseCli, "CTL", os.ExpandEnv("$HOME/.countercli"))
+	cmd := cli.PrepareMainCmd(GaiaCli, "CTL", os.ExpandEnv("$HOME/.countercli"))
 	cmd.Execute()
 }
